Create game logiv
This is my first ever real commit. Take it easy on me.
I'm addressing the current state of the game as per
educational purposes. Currently trying to mess around
with WSL, Windows and Xubuntu running on a virtual
machine in order to figure out the most effective ways
to geta round programming as well as develop my self
skills.